fix: apply blur in stellarium web

+++ telescope/Slider/slider.js
@@ -1,6 +1,6 @@
 
 blurSlider.addEventListener('input', () => {
-    currentBlur = parseInt(blurSlider.value);
+    currentBlur = parseFloat(blurSlider.value);
 
     updateDisplayBlur();
 });
